refactor: Create member_list_item jsx & scss
