+++ app/assets/stylesheets/components/member_list_item.scss
@@ -1,0 +1,48 @@
+.member-list-item {
+  box-sizing: border-box;
+  padding-bottom: 20px;
+  > div {
+    width: 133px;
+    height: 200px;
+    border-radius: 8px;
+    background-color: white;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    transition: box-shadow 0.25s cubic-bezier(0.4, 0, 0.2, 1),
+      transform 0.25s cubic-bezier(0.4, 0, 0.2, 1),
+      -webkit-box-shadow 0.25s cubic-bezier(0.4, 0, 0.2, 1),
+      -webkit-transform 0.25s cubic-bezier(0.4, 0, 0.2, 1);
+    > h3 {
+      padding-top: 20px;
+      text-align: center;
+      font-size: 14px;
+      font-weight: 500;
+      line-height: 1.2;
+    }
+    > p {
+      padding-top: 2px;
+      font-size: 12px;
+      color: rgba(46, 62, 72, 0.6);
+      line-height: 1.45;
+    }
+  }
+  > div:hover {
+    box-shadow: 0 0 4px 0 rgba(46, 62, 72, 0.12),
+      0 4px 12px 0 rgba(46, 62, 72, 0.12);
+    transform: translateY(-1px);
+  }
+  img,
+  i {
+    height: 72px;
+    width: 72px;
+    font-size: 72px;
+    margin-top: 20px;
+  }
+  img {
+    border-radius: 999px;
+  }
+}
+.mli-pad-right {
+  padding-right: 20px;
+}
