+++ frontend/components/member_list.jsx
@@ -1,4 +1,5 @@
 import React from "react";
+import MemberListItem from "./member_list_item";
 
 const MemberList = ({ header, members, isMember }) => {
   members.sort((memberA, memberB) => {
@@ -27,44 +28,6 @@ const MemberList = ({ header, members, isMember }) => {
         })}
       </ul>
     </section>
-  );
-};
-
-const firstLast = name => {
-  const nameArr = name.split(" ");
-  const firstName = nameArr[0];
-  const lastName = nameArr.length === 1 ? null : nameArr[nameArr.length - 1];
-  return [firstName, lastName];
-};
-
-const MemberListItem = ({ member, shouldPadRight, isMember }) => {
-  const avatar = member.avatarUrl ? (
-    <img src={member.avatarUrl} />
-  ) : (
-    <i className="far fa-user-circle" />
-  );
-  const klass = `member-list-item${shouldPadRight ? " mli-pad-right" : ""}`;
-  const [firstName, lastName] = firstLast(member.name);
-  const diff = isMember? -1 : 0;
-  const sharedGroups = member.common_group_counts+diff;
-  const groupOrGroups = sharedGroups === 1 ? "group" : "groups";
-  return (
-    <li className={klass}>
-      <div>
-        {avatar}
-        <h3>
-          {firstName}
-          {lastName ? <br /> : null}
-          {lastName}
-        </h3>
-        <p> {member.memberType} </p>
-        {member.common_group_counts ? (
-          <p>
-            {sharedGroups} shared {groupOrGroups}
-          </p>
-        ) : null}
-      </div>
-    </li>
   );
 };
